added logs for header
